Ball Screw Short BNT Brand + Info
A short ball screw generated through the BNT website along with the
bearing, This also includes an assembly of both the ball screw and the
bearing which Ross pointed out will be roughly the same as the one that
we choose. This also contains a file which has other information
regarding ball screws and bearings ranging from different sizes, thread
sizes and other.